Moved client thread and secure_importer to different files to alleviate space in engine file
